Rename requests directory
Importing from this directory was not working because it shadowed the name of the requests python module
