+++ request_definitions/jabber.py
