+++ request_definitions/trend.py
