+++ request_definitions/xbox.py
